Update PlanEntryRepetition.cs
now add interval-1 day at first to have the last day of the interval and at the and the missing day is added to start the next interval on the correct date.
if the end of the interval is past the end date it will be set to the end date instead.